implement rendering and default config with example image

+++ src/main/java/cz/lukaskabc/minecraft/mod_loader/loading/simple_custom_early_loading/elements/ProgressBar.java
@@ -76,6 +76,6 @@ public class ProgressBar {
     }
 
     public interface BarPosition {
-        int[] location(RenderElement.DisplayContext context);
+        float[] location(RenderElement.DisplayContext context);
     }
 }
